requirement.txt
requests
numpy
pandas
matplotlib
tensorflow==2.19.0
joblib
paho-mqtt
scikit-learn
